+++ tests/testthat/test_plotPairs.R
@@ -52,6 +52,12 @@ test_that("plotPairs fails with incorrect arguments", {
     expect_error(plotPairs(se = se, selAssay = "counts", colorByCorrelation = "TRUE"))
     expect_error(plotPairs(se = se, selAssay = "counts", colorByCorrelation = c(TRUE, FALSE)))
     
+    expect_error(plotPairs(se = se, selAssay = "counts", corrColorRange = 1))
+    expect_error(plotPairs(se = se, selAssay = "counts", corrColorRange = c(-1, 0.5)))
+    expect_error(plotPairs(se = se, selAssay = "counts", corrColorRange = c(1, 0.5)))
+    expect_error(plotPairs(se = se, selAssay = "counts", corrColorRange = c("0.5", "1")))
+    expect_error(plotPairs(se = se, selAssay = "counts", corrColorRange = c(0, 2)))
+    
     expect_error(plotPairs(se = se, selAssay = "counts", addIdentityLine = 1))
     expect_error(plotPairs(se = se, selAssay = "counts", addIdentityLine = "TRUE"))
     expect_error(plotPairs(se = se, selAssay = "counts", addIdentityLine = c(TRUE, FALSE)))
@@ -63,7 +69,7 @@ test_that("plotPairs works with correct arguments", {
                               pseudocount = 1, corMethod = "pearson",
                               histBreaks = 40, pointsType = "points", corSizeMult = 5,
                               corSizeAdd = 2, pointSize = 0.1, pointAlpha = 0.3,
-                              colorByCorrelation = TRUE, 
+                              colorByCorrelation = TRUE, corrColorRange = NULL,
                               addIdentityLine = FALSE), "ggmatrix")
 
     ## Not log-transformed
@@ -71,7 +77,7 @@ test_that("plotPairs works with correct arguments", {
                               pseudocount = 1, corMethod = "pearson",
                               histBreaks = 40, pointsType = "points", corSizeMult = 5,
                               corSizeAdd = 2, pointSize = 0.1, pointAlpha = 0.3,
-                              colorByCorrelation = TRUE, 
+                              colorByCorrelation = TRUE, corrColorRange = NULL, 
                               addIdentityLine = FALSE), "ggmatrix")
 
     ## No pseudocount
@@ -79,7 +85,7 @@ test_that("plotPairs works with correct arguments", {
                               pseudocount = 0, corMethod = "pearson",
                               histBreaks = 40, pointsType = "points", corSizeMult = 5,
                               corSizeAdd = 2, pointSize = 0.1, pointAlpha = 0.3,
-                              colorByCorrelation = TRUE, 
+                              colorByCorrelation = TRUE, corrColorRange = NULL, 
                               addIdentityLine = FALSE), "ggmatrix")
 
     ## Spearman correlation
@@ -87,7 +93,7 @@ test_that("plotPairs works with correct arguments", {
                               pseudocount = 1, corMethod = "spearman",
                               histBreaks = 40, pointsType = "points", corSizeMult = 5,
                               corSizeAdd = 2, pointSize = 0.1, pointAlpha = 0.3,
-                              colorByCorrelation = TRUE, 
+                              colorByCorrelation = TRUE, corrColorRange = NULL, 
                               addIdentityLine = FALSE), "ggmatrix")
 
     ## Fewer breaks
@@ -95,7 +101,7 @@ test_that("plotPairs works with correct arguments", {
                               pseudocount = 1, corMethod = "pearson",
                               histBreaks = 5, pointsType = "points", corSizeMult = 5,
                               corSizeAdd = 2, pointSize = 0.1, pointAlpha = 0.3,
-                              colorByCorrelation = TRUE, 
+                              colorByCorrelation = TRUE, corrColorRange = NULL, 
                               addIdentityLine = FALSE), "ggmatrix")
 
     ## smoothScatter
@@ -103,7 +109,7 @@ test_that("plotPairs works with correct arguments", {
                               pseudocount = 1, corMethod = "pearson",
                               histBreaks = 40, pointsType = "smoothscatter", corSizeMult = 5,
                               corSizeAdd = 2, pointSize = 0.1, pointAlpha = 0.3,
-                              colorByCorrelation = TRUE, 
+                              colorByCorrelation = TRUE, corrColorRange = NULL, 
                               addIdentityLine = FALSE), "ggmatrix")
 
     ## Change font size to correlation relation
@@ -111,7 +117,7 @@ test_that("plotPairs works with correct arguments", {
                               pseudocount = 1, corMethod = "pearson",
                               histBreaks = 40, pointsType = "points", corSizeMult = 2,
                               corSizeAdd = 0, pointSize = 0.1, pointAlpha = 0.3,
-                              colorByCorrelation = TRUE, 
+                              colorByCorrelation = TRUE, corrColorRange = NULL, 
                               addIdentityLine = FALSE), "ggmatrix")
 
     ## Point size, alpha
@@ -119,7 +125,7 @@ test_that("plotPairs works with correct arguments", {
                               pseudocount = 1, corMethod = "pearson",
                               histBreaks = 40, pointsType = "points", corSizeMult = 5,
                               corSizeAdd = 2, pointSize = 1, pointAlpha = 3,
-                              colorByCorrelation = TRUE, 
+                              colorByCorrelation = TRUE, corrColorRange = NULL, 
                               addIdentityLine = FALSE), "ggmatrix")
 
     ## Don't color by correlation
@@ -127,7 +133,15 @@ test_that("plotPairs works with correct arguments", {
                               pseudocount = 1, corMethod = "pearson",
                               histBreaks = 40, pointsType = "points", corSizeMult = 5,
                               corSizeAdd = 2, pointSize = 0.1, pointAlpha = 0.3,
-                              colorByCorrelation = FALSE, 
+                              colorByCorrelation = FALSE, corrColorRange = NULL, 
+                              addIdentityLine = FALSE), "ggmatrix")
+    
+    ## Change color range
+    expect_s3_class(plotPairs(se = se, selAssay = "counts", doLog = TRUE,
+                              pseudocount = 1, corMethod = "pearson",
+                              histBreaks = 40, pointsType = "points", corSizeMult = 5,
+                              corSizeAdd = 2, pointSize = 0.1, pointAlpha = 0.3,
+                              colorByCorrelation = TRUE, corrColorRange = c(0.7, 1), 
                               addIdentityLine = FALSE), "ggmatrix")
     
     ## Add identity line
@@ -135,7 +149,7 @@ test_that("plotPairs works with correct arguments", {
                               pseudocount = 1, corMethod = "pearson",
                               histBreaks = 40, pointsType = "points", corSizeMult = 5,
                               corSizeAdd = 2, pointSize = 0.1, pointAlpha = 0.3,
-                              colorByCorrelation = TRUE, 
+                              colorByCorrelation = TRUE, corrColorRange = NULL, 
                               addIdentityLine = TRUE), "ggmatrix")
 
     ## Make one correlation negative
@@ -146,5 +160,6 @@ test_that("plotPairs works with correct arguments", {
                               pseudocount = 1, corMethod = "pearson",
                               histBreaks = 40, pointsType = "points", corSizeMult = 5,
                               corSizeAdd = 2, pointSize = 0.1, pointAlpha = 0.3,
-                              colorByCorrelation = TRUE), "ggmatrix")
+                              colorByCorrelation = TRUE, corrColorRange = NULL,
+                              addIdentityLine = FALSE), "ggmatrix")
 })